+++ tests/TestCase.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Bigbank\DigiDoc\Test;
+
+abstract class TestCase extends \PHPUnit_Framework_TestCase
+{
+
+}